+++ brain/circuit.cpp
@@ -25,6 +25,7 @@
 
 #include "synapses.h"
 #include "synapsesStream.h"
+#include "types.h"
 
 #include <bbp/sonata/edges.h>
 
@@ -49,17 +50,22 @@ URIs _getMorphologyURIs(const Circuit::Impl& circuit, const GIDSet& gids)
 
 Matrix4fs _getTransforms(const Circuit::Impl& circuit, const GIDSet& gids)
 {
-    const Vector3fs& positions = circuit.getPositions(gids);
-    const Quaternionfs& rotations = circuit.getRotations(gids);
+    const Vector3fs positions = circuit.getPositions(gids);
+    const Quaternionfs rotations = circuit.getRotations(gids);
     if (positions.size() != rotations.size())
         throw std::runtime_error(
             "Positions not equal rotations for given GIDs");
 
     Matrix4fs transforms(positions.size());
 
-#pragma omp parallel for
+    #pragma omp parallel for
     for (size_t i = 0; i < positions.size(); ++i)
-        transforms[i] = Matrix4f(rotations[i], positions[i]);
+    {
+        transforms[i] = glm::mat4_cast(rotations[i]);
+        transforms[i][3] = glm::vec4(positions[i].x, positions[i].y,
+                                     positions[i].z, 1.0);
+    }
+        
     return transforms;
 }
 
@@ -275,13 +281,13 @@ private:
         const size_t end = somaID == sections.size() - 1
                                ? points.size()
                                : sections[somaID + 1][0];
-        Vector3f centroid;
+        glm::vec3 centroid;
         for (size_t i = start; i != end; ++i)
-            centroid += points[i].get_sub_vector<3, 0>();
+            centroid += glm::vec3(points[i]);
         centroid /= float(end - start);
 
         // Translating all points
-        if (centroid.length() < 1e-6)
+        if (glm::length(centroid) < 1e-6)
             return; // Do not recenter is almost there
 #pragma omp parallel for
         for (size_t i = 0; i < points.size(); ++i)