+++ brain/detail/circuit.h
@@ -96,15 +96,18 @@ void _assign(const ::MVD3::Range& range, const GIDSet& gids, SrcArray& src,
         *j++ = assignOp(src[gid - range.offset - 1]);
 }
 
-Vector3f _toVector3f(const ::MVD3::Positions::const_subarray<1>::type& subarray)
+glm::vec3 _toVector3f(const ::MVD3::Positions::const_subarray<1>::type& subarray)
 {
-    return Vector3f(subarray[0], subarray[1], subarray[2]);
+    return glm::vec3(subarray[0], subarray[1], subarray[2]);
 }
 
-Quaternionf _toQuaternion(
+glm::quat _toQuaternion(
     const ::MVD3::Rotations::const_subarray<1>::type& subarray)
 {
-    return Quaternionf(subarray[0], subarray[1], subarray[2], subarray[3]);
+    return glm::quat(static_cast<float>(subarray[0]), 
+                     static_cast<float>(subarray[1]), 
+                     static_cast<float>(subarray[2]), 
+                     static_cast<float>(subarray[3]));
 }
 
 template <typename T>
@@ -331,7 +334,7 @@ public:
             const float y = attributeY[gid - startIdx];
             const float z = attributeZ[gid - startIdx];
 
-            output.push_back(Vector3f(x, y, z));
+            output.push_back(glm::vec3(x, y, z));
         }
         return output;
     }
@@ -388,7 +391,7 @@ public:
             // Special case for missing rotation angles.
             if (rX == 0 && rY == 0 && rZ == 0)
             {
-                output.push_back(Quaternionf());
+                output.push_back(glm::quat_cast(glm::mat4(1.f)));
                 continue;
             }
 
@@ -398,8 +401,12 @@ public:
             // These are the values given by multiplying the rotation
             // matrices for R(X)R(Y)R(Z) i.e. extrinsic rotation around Z
             // then Y then X
-            vmml::Matrix3f mtx;
+            glm::mat3 mtx (1.f);
 
+            mtx[0] = glm::vec3(cY * cZ, cZ * sX * sY + cX * sZ, sX * sZ - cX * cZ * sY);
+            mtx[1] = glm::vec3(-cY * sZ, cX * cZ - sX * sY * sZ, cZ * sX + cX * sY * sZ);
+            mtx[2] = glm::vec3(sY, -cY * sX, cX * cY);
+            /*
             mtx(0, 0) = cY * cZ;
             mtx(0, 1) = -cY * sZ;
             mtx(0, 2) = sY;
@@ -409,8 +416,8 @@ public:
             mtx(2, 0) = sX * sZ - cX * cZ * sY;
             mtx(2, 1) = cZ * sX + cX * sY * sZ;
             mtx(2, 2) = cX * cY;
-
-            output.push_back(Quaternionf(mtx));
+            */
+            output.push_back(glm::quat_cast(mtx));
         }
 
         return output;
@@ -706,7 +713,7 @@ public:
         if (gids.empty())
             return Vector3fs();
 
-        const brion::NeuronMatrix& data =
+        const brion::NeuronMatrix data =
             _circuit.get(gids, brion::NEURON_POSITION_X |
                                    brion::NEURON_POSITION_Y |
                                    brion::NEURON_POSITION_Z);
@@ -717,9 +724,9 @@ public:
         {
             try
             {
-                positions[i] = brion::Vector3f(lexical_cast<float>(data[i][0]),
-                                               lexical_cast<float>(data[i][1]),
-                                               lexical_cast<float>(data[i][2]));
+                positions[i] = glm::vec3(lexical_cast<float>(data[i][0]),
+                                         lexical_cast<float>(data[i][1]),
+                                         lexical_cast<float>(data[i][2]));
             }
             catch (const boost::bad_lexical_cast&)
             {
@@ -737,7 +744,7 @@ public:
         if (gids.empty())
             return size_ts();
 
-        const brion::NeuronMatrix& matrix =
+        const brion::NeuronMatrix matrix =
             _circuit.get(gids, brion::NEURON_MTYPE);
         size_ts result(matrix.shape()[0]);
 
@@ -758,7 +765,7 @@ public:
         if (gids.empty())
             return size_ts();
 
-        const brion::NeuronMatrix& matrix =
+        const brion::NeuronMatrix matrix =
             _circuit.get(gids, brion::NEURON_ETYPE);
         size_ts result(matrix.shape()[0]);
 
@@ -780,7 +787,7 @@ public:
             return Quaternionfs();
 
         const float deg2rad = float(M_PI) / 180.f;
-        const brion::NeuronMatrix& data =
+        const brion::NeuronMatrix data =
             _circuit.get(gids, brion::NEURON_ROTATION);
         Quaternionfs rotations(gids.size());
 
@@ -791,7 +798,7 @@ public:
             {
                 // transform rotation Y angle in degree into rotation quaternion
                 const float angle = lexical_cast<float>(data[i][0]) * deg2rad;
-                rotations[i] = Quaternionf(angle, Vector3f(0, 1, 0));
+                rotations[i] = glm::angleAxis(angle, glm::vec3(0.f, 1.f, 0.f));
             }
             catch (const boost::bad_lexical_cast&)
             {
@@ -809,7 +816,7 @@ public:
         if (gids.empty())
             return Strings();
 
-        const brion::NeuronMatrix& matrix =
+        const brion::NeuronMatrix matrix =
             _circuit.get(gids, brion::NEURON_MORPHOLOGY_NAME);
         Strings result;
         result.reserve(matrix.shape()[0]);